Remove proxy node functionality
When creating mocks on the test cluster, we realized it was hard to call
back to the original test manager node via the Braid proxy. Instead of
making complex functionality to proxy backwards via the Braid proxy
node, we opted for removing the proxy node completely. This means that
you _should_ start the test manager node as hidden, to avoid it becoming
part of the test cluster.