Increase stopping timeout to handle many nodes
Waiting for all nodes to exit takes some time because of init:stop()
running on all nodes. This is a temprary increase in the request timeout
to allow for a slightly larger number of nodes to be stopped.

However, this is not a solution. Ideally, the timeout should depend on
the number of nodes or somehow be handler on a per-node basis. Needs
further investigation.